pyproject.toml
[project]
name = "slack-emoji-pipeline"
version = "1.0.0"
description = "Professional 3-stage automated pipeline for high-fidelity Slack emojis."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "playwright>=1.40.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
